requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.2.2
frozenlist==1.4.1
greenlet==3.0.3
grpcio==1.65.4
grpcio-tools==1.65.4
h11==0.14.0
h2==4.1.0
hpack==4.0.0
htmltools==0.5.3
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.12
langchain-community==0.2.11
langchain-core==0.2.29
langchain-openai==0.1.21
langchain-openai
langchain-text-splitters==0.2.2
langsmith==0.1.98
linkify-it-py==2.0.3
lxml==5.3.0
markdown-it-py==3.0.0
marshmallow==3.21.3
mdit-py-plugins==0.4.1
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.40.3
orjson==3.10.7
packaging==24.1
pillow==10.4.0
portalocker==2.10.1
prompt-toolkit==3.0.36
protobuf==5.27.3
pydantic==2.8.2
pydantic_core==2.20.1
pypdf==4.3.1
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv
python-multipart==0.0.9
python-pptx==1.0.2
PyYAML==6.0.2
qdrant-client==1.10.1
questionary==2.0.1
regex==2024.7.24
requests==2.32.3
shiny==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
uc-micro-py==1.0.3
unstructured==0.15.8
unstructured-ingest
urllib3==2.2.2
uvicorn==0.30.4
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
XlsxWriter==3.2.0
yarl==1.9.4
